refactor: move specific result store functions from utils to calc
